requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.6.0
anyio==4.3.0
attrs==18.2.0
Babel==2.14.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
cryptography==43.0.1
databases==0.9.0
dnspython==2.6.1
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.8
email_validator==2.1.1
et-xmlfile==1.1.0
etelemetry==0.3.1
fastapi==0.110.2
filelock==3.15.4
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
looseversion==1.3.0
lxml==5.2.1
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.0.5
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nodeenv==1.9.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.1
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdflatex==0.1.3
pdfrw==0.4
pillow==10.3.0
pip-autoremove==0.10.0
prisma==0.13.1
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pydot==2.0.0
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypandoc==1.13
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
reportlab==4.2.2
requests==2.32.3
rsa==4.9
scipy==1.14.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.37.2
tomlkit==0.13.0
tqdm==4.66.4
traits==6.3.2
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.29.0
waitress==3.0.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
zipfile2==0.0.12
